Move tls::Connection to a submodule.
