+++ proxy/src/transport/tls/connection.rs
@@ -1,0 +1,78 @@
+use std::io;
+use std::net::SocketAddr;
+
+use bytes::Buf;
+use futures::Future;
+use tokio::prelude::*;
+use tokio::net::TcpStream;
+
+use transport::{AddrInfo, Io};
+
+use super::{
+    rustls::ServerSession,
+    tokio_rustls::{ServerConfigExt, TlsStream},
+
+    ServerConfig,
+};
+
+// In theory we could replace `TcpStream` with `Io`. However, it is likely that
+// in the future we'll need to do things specific to `TcpStream`, so optimize
+// for that unless/until there is some benefit to doing otherwise.
+#[derive(Debug)]
+pub struct Connection(TlsStream<TcpStream, ServerSession>);
+
+impl Connection {
+    pub fn accept(socket: TcpStream, ServerConfig(config): ServerConfig)
+                  -> impl Future<Item = Connection, Error = io::Error>
+    {
+        config.accept_async(socket).map(Connection)
+    }
+}
+
+impl io::Read for Connection {
+    fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
+        self.0.read(buf)
+    }
+}
+
+impl AsyncRead for Connection {
+    unsafe fn prepare_uninitialized_buffer(&self, buf: &mut [u8]) -> bool {
+        self.0.prepare_uninitialized_buffer(buf)
+    }
+}
+
+impl io::Write for Connection {
+    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
+        self.0.write(buf)
+    }
+
+    fn flush(&mut self) -> io::Result<()> {
+        self.0.flush()
+    }
+}
+
+impl AsyncWrite for Connection {
+    fn shutdown(&mut self) -> Poll<(), io::Error> {
+        self.0.shutdown()
+    }
+
+    fn write_buf<B: Buf>(&mut self, buf: &mut B) -> Poll<usize, io::Error> {
+        self.0.write_buf(buf)
+    }
+}
+
+impl AddrInfo for Connection {
+    fn local_addr(&self) -> Result<SocketAddr, io::Error> {
+        self.0.get_ref().0.local_addr()
+    }
+
+    fn get_original_dst(&self) -> Option<SocketAddr> {
+        self.0.get_ref().0.get_original_dst()
+    }
+}
+
+impl Io for Connection {
+    fn shutdown_write(&mut self) -> Result<(), io::Error> {
+        self.0.get_mut().0.shutdown_write()
+    }
+}
